refactor(self-review): share the prose extractor, keep the sentence splitters apart
check_aphorism_density and check_rhetorical_density measure different tells over the SAME text —
body prose with front matter, headings, tables, quotes, fences, list items, citations and inline
markup removed — and each carried its own byte-identical copy of that extractor. Two copies of one
definition is one drift away from two detectors reporting densities over different denominators
while both claim to describe the same prose. The extractor and its three regexes move to a shared
same-directory _prose.py, the pattern _frontmatter.py already established.

What is NOT shared, deliberately: sentence splitting. The independent review that prompted this
called all three prose detectors duplicative; reading them says otherwise.

    aphorism    (?<=[.!?])\s+(?=[A-Z"“])          sentences: >= 2 words
    rhetorical  (?<=[.!?])\s+(?=["“(]?[A-Z0-9])   sentences: any non-empty

The rhetorical splitter starts a sentence at a digit or an opening parenthesis; the aphorism one
does not, and it drops one-word fragments so a run of them cannot masquerade as clipped
declaratives. Unifying those would have silently moved every density in both detectors — a
behaviour change wearing a refactor's clothes. Each keeps its own splitter and sentence filter.

humanize/check_sentence_variety keeps its own extractor: skills are self-contained and cross-skill
imports are forbidden. That is architecture, not duplication.

The shared module was extracted programmatically from the source, not retyped — the first draft was
hand-copied and three definitions came out subtly different (`[*_`]` vs `[*_`]+`, a reordered
character class), which is exactly the failure the extraction was meant to prevent.

Proven inert: ten detector outputs across five fixtures are byte-identical before and after; both
test suites pass; the detector count is unchanged at 84.
